pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[project]
name = "cityvision"
version = "0.1.0"
description="CityVision Video Analytics"
authors = [
  {name = "Sahand Somi", email = "sahand.somi@edmonton.ca"}
]
maintainers = [
  {name = "Sahand Somi", email = "sahand.somi@edmonton.ca"}
]

license = {text = "AGPL-3.0"}

dependencies = [
    "ultralytics==8.2.83",
    "opencv-python==4.10.0.84",
    "numpy==1.24.4",
    "google-cloud-storage==2.18.2",
    "google-cloud-bigquery==3.25.0",
    "google-cloud-pubsub==2.23.0",
    "ndjson==0.3.1",
    "pandas==2.2.1",
    "lapx==0.5.10",
    "onnx==1.16.2",
    "supervision==0.23.0"
]